refactor: count에 대한 상태를 Page에서 관리한다
