+++ src/components/App.jsx
@@ -1,24 +1,9 @@
-import React, { useState } from 'react';
+import React from 'react';
 
 import Page from './Page';
 
 export default function App() {
-  const [state, setState] = useState({
-    count: 0,
-  });
-
-  const { count } = state;
-
-  function handleClick() {
-    setState({
-      count: count + 1,
-    });
-  }
-
   return (
-    <Page
-      count={count}
-      onClick={handleClick}
-    />
+    <Page />
   );
 }
